azure-function/requirements.txt
azure-functions
pydantic
requests
python-telegram-bot